Add HACKING file with 'Philosophy' paragraph from the Userbase page
https://userbase.kde.org/Dolphin has a 'Philosophy' paragraph (which
used to be a page on dolphin.kde.org) explaining when a new feature
should be added to dolphin. Userbase is a wiki aimed to actual user of
our applications, while this information is aimed to dolphin developers
who want to add a new feature to dolphin. Thus it is more appropriate to
have this paragraph in the git repo.