Fix missing update of file name label in information panel without preview
Summary:
The file name label was only updated when the preview was shown, after
the function was moved into a conditional section. This is a regression
caused by commit 405aa4878f75 ("[InformationPanel] Hide the video when the
preview is disabled, avoid computing the preview when it is disabled").

Test Plan:
Hide the preview in dolphin's information panel
Hover different files, the name label reflects the hovered file again

Reviewers: #dolphin, elvisangelaccio, meven, ngraham

Reviewed By: meven

Subscribers: kfm-devel

Tags: #dolphin

Differential Revision: https://phabricator.kde.org/D20660