Do not match daemonized dolphin instances: part 2
Summary:
`dolphin --daemon` does not have the `/dolphin/Dolphin_1` dbus path,
because it doesn't have any DolphinMainWindow.

Checking if the QDBusInterface is valid is not enough, we also need to look for dbus errors.

This should have been part of commit 7df5fc8edb54153f. Not sure how I missed it.

BUG: 408244
FIXED-IN: 19.08.0

Test Plan:
1. Make sure the new "Open new folders in tabs" setting is enabled.
2. Run `dolphin --deamon` and make sure is the only dolphin process around.
3. Call the `org.freedesktop.FileManager1.ShowFolders` method from QDbusViewer.

Subscribers: kfm-devel

Tags: #dolphin

Differential Revision: https://phabricator.kde.org/D23087