Unbreak context menu showing when right clicking on an empty area
In D22149, a request was made to remove a variable assignment ("action",
on line 389 in src/dolphincontextmenu.cpp. However the change went too
far, and it actually removed the exec() call, not just the assignment
which should have stayed.

This resulted in the context menu not being shown at all.

Adding back exec(m_pos) at least allows for the context menu to show.

CCMAIL: nate@kde.org