[dolphin] Animate gifs on preview
Summary:
Adds the capability to view animated images on the preview in the information panel.
This was a request from a user back in 2009 (https://bugs.kde.org/show_bug.cgi?id=182257), but I think nowadays with stickers/memes and what not, it's even more useful.
It keeps the size default transition of the preview viewer before starting the animation, so that the visual integration is smoother.

{F7289110}

FEATURE: 182257
FIXED-IN: 19.11.80

Test Plan: Open the information panel and hover over some animated images (gif/webp/mng)

Reviewers: #dolphin, #vdg, ngraham, elvisangelaccio

Reviewed By: #vdg, ngraham

Subscribers: pino, fuksitter, meven, broulik, kfm-devel

Tags: #dolphin

Differential Revision: https://phabricator.kde.org/D23538