Update searchbox on URL changes (Fix D24369)
Summary:
In D24369 I introduced a regression: Where now the searchbox is updated on url changes for search urls, it is not updated anymore for regular folders. The effect is that not-indexed folders can have the extra options enabled or vice-versa. This patch fixes it, moving back the updating logic to `setSearchModeEnable`.

Test Plan: - Searchbox updates properly when changing between search urls and regular folders, either indexed or nor-indexed.

Reviewers: #dolphin, elvisangelaccio

Subscribers: kfm-devel, elvisangelaccio

Tags: #dolphin

Differential Revision: https://phabricator.kde.org/D24376