Hide search extra options when not available
Summary:
Hide the search extra options when they are not available, because the current location is not indexed or Baloo is not enabled.
The button is disabled and shows a tooltip to provide feedback of why it's not available.
Depends on D24478

{F7531893}

CCBUG: 396898
FEATURE: 318580
FEATURE: 408680
FIXED-IN: 19.11.80

Test Plan:
 - `Ctrl-F` on a non-indexed location
 - `balooctl disable` and `Ctrl-F` on an indexed location

On both cases the extra options are not shown, the button is disabled and the reason is shown in the tooltip
The user selection is remembered between indexed locations.

Reviewers: #dolphin, elvisangelaccio, meven, ngraham

Reviewed By: meven

Subscribers: kfm-devel

Tags: #dolphin

Differential Revision: https://phabricator.kde.org/D24446