New design for extended search options
Summary:
Replace the arrays of radiobuttons in dolphin search box to more simple dropdown boxes.
This provides a leaner look, saving a lot user space on the main view and it is more consistent with the 'Search tools' in the most known sites (Google, DuckDuckGo, etc.)
There is room for improvement, as `QComboBox` doesn't match perfectly with the tool buttons used avobe, but I think it is an improvement over the current situation.

BEFORE:
{F7581883}

AFTER:
{F7580952}

Test Plan: Same behavior with different aesthetics

Reviewers: elvisangelaccio, #dolphin, #vdg

Subscribers: ndavis, ngraham, kfm-devel

Tags: #dolphin

Differential Revision: https://phabricator.kde.org/D24602