Use more conventional "Add to Places" text for that action
Summary:
I originally implemented this to show "Add <current folder> to Places" to make the menu
item more clear when invoked from the file menu or the viewport context menu because I
thought it wasn't totally clear that the action would be referring to the folder for the
visible view.

However I think in retrospect that this was a mistake. No other items in the file menu
or the viewport context menu name the current folder like this, so clearly there wasn't
a problem. And then the item's text doesn't match the text for other items.

Thie patch removes that behavior and makes the name match the style of all the other menu
item text.

BUG: 416064
FIXED-IN: 20.04.0

Reviewers: #dolphin, #vdg, meven

Reviewed By: #dolphin, meven

Subscribers: elvisangelaccio, meven, kfm-devel

Tags: #dolphin

Differential Revision: https://phabricator.kde.org/D26590