Use same logic for "no extension" case with Duplicate feature
Summary:
In the "no extension" case, we weren't separating out the path and the original filename,
breaking the feature for languages where the word "copy" would be at the beginning of the
filename, not after it (e.g. "copia de foo" in Spanish, and similar in other romance
languages). This patch fixes that by separating the original path and filename in the no
extension case as is done for the other case, which should solve the issue.

BUG: 419070
FIXED-IN: 20.04.0

Test Plan:
No changes in English; should fix the issue in Spanish once new translations are done
(see https://bugs.kde.org/show_bug.cgi?id=419070 for details)

Reviewers: #dolphin, elvisangelaccio, arojas, meven, pino, #localization

Reviewed By: #dolphin, elvisangelaccio

Subscribers: aacid, kfm-devel

Tags: #dolphin

Differential Revision: https://phabricator.kde.org/D28227